adding final calls to item mean model based recommender
