+++ non_personalized/item_mean_model_provider.py
@@ -12,14 +12,8 @@ class ItemMeanModelProvider(ItemMeanModel):
             raise TypeError("unkown data type")
         else:
             self.__ratings_data=ratings_data
-    def compute_mean(self):
+    def __compute_mean(self):
         means_df = self.ratings_data.groupby(["itemID"])["rating"].mean()
         return means_df.to_dict()
     def get(self):
-        return ItemMeanModel(self.compute_mean())
-if __name__ == "__main__":
-    ratings = pd.read_csv("data/ratings.csv")
-    del ratings["timestamp"]
-    ratings.columns=["userID","itemID","rating"]
-    IMMP = ItemMeanModelProvider(ratings)
-    print(IMMP.get().get_mean_rating(17))
+        return ItemMeanModel(self.__compute_mean())
